bug fix: remove no-op fp32 cast

+++ model.py
@@ -33,7 +33,7 @@ class TransformerDecoder(nnx.Module):
 
         # project back to vocabulary
         h = self.out_ln(h)
-        logits = self.token_embed_out.attend(h.astype(jnp.float32)) # [B, T, V]
+        logits = self.token_embed_out.attend(h) # [B, T, V]
         logits = jax.lax.with_sharding_constraint(logits, P('data', None, 'model'))
         return logits
